Write attributions and valuation only at the end after processing
This was causing a bug formerly where the original valuation was being
passed in processing each new payment, even if the previous payment
may have inflated the valuation. As a result, at most one
payment (i.e. the last processed) was reflected in the written
inflated valuation rather than all of them.