Add DIA allocation and renormalize
We're not adding the Old Abe allocation here since it would cause a
pointless accounting loop of recording payouts and payments and
transactions for that 1% that is ultimately just coming to Old Abe
itself.

This also marks attributions as dilutable or not, as we have been
doing in other projects for this purpose.