fix test expectations for unpayable contributor
The tests still fail, but the expectations are correct and the actual
values produced are also likely correct, except that they reflect a
higher precision than the level we expect to find in actual accounting
files, that is, of dollars and cents (or any currency units).

The handling of "dust" will inform how we go from the internal
high-precision setting to the ultimate, low-precision, written
amounts, so, leaving off fixing these tests / the underlying code
for now, until such a time as we decide on the handling of dust.