Only commit after both money_in and money_out have run
This ensures that we do not run into the problem where interrupted
jobs end up trying to commit the same changes to accounting records
more than once. Also, ensure that we do not attempt to run money_out
if money_in fails. Finally, note a potential race condition and ways
to handle it.